+++ ui/components/multichain/network-manager/network-manager.stories.tsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import configureStore from '../../../store/store';
 import testData from '../../../../.storybook/test-data';
 
-import { NetworkManager } from '.';
+import { NetworkManager } from './network-manager';
 
 const store = configureStore(testData);
